phew what a day 2

+++ Game/Code/Game/Entity/Adversary/Components/AdversaryStatus.cs
@@ -6,9 +6,9 @@ namespace Daikon.Game;
 public partial class AdversaryStatus : EntityStatus
 {
     [Signal]
-    public delegate void ThreatInflictedEventHandler(float threat, Entity entity);
+    public delegate void ThreatInflictedEventHandler(int threat, Entity entity);
 
-    public void InflictThreat(float threat, Entity entity) 
+    public void InflictThreat(int threat, Entity entity) 
     {
         EmitSignal(SignalName.ThreatInflicted, threat, entity);
     }
